Added an option to specify the default width of splits
A lot of the time I just want the terminal to be there, even if I am not actively looking at it. With the default split width being 50% it was kind of obnoxious unless I `vertical resize 80` every time things move. This adds a new `setup` function with lets you override default options, including a `split_default_width` option which lets you directly set the number of columns.