When restarting a job from the main menu it no longer replaces the buffer in a window that previously was a projectlaunch terminal buffer but isn't anymore
If you opened a split window, then changed the buffer to a file or something, then hit `R` on a job in the main menu it'd restart the job and replace the previous split window with the new job's terminal, even though that window no longer "belongs" to projectlaunch. This reuses the new `is_split_usable()`. I had to rearrange term.lua's `M.restart_job` so it'd check the active window type before replacing the old job in the jobs table, otherwise it'd always think the old job's terminal buffer isn't a projectlaunch terminal buffer and it would break the `restart_command_in_split` method.