Add a way to open the launch menu when nvim starts
This adds a couple new methods that are exported, one to toggle the launch menu (some minor refactoring had to be done to make this work like the main menu's toggle function), and a way to check if there are any commands that can be run. This way you can write an autocmd to open the launch menu automatically when opening nvim.